Basics of Command Based Driving, some setup
Not really command based, but the Idea.

Get Joysticks in teleop, send to global, in control loop grab it and convert to motor values, and write motors.

Initialize Subsystem in the DriveLoop so you can see it and its all together, but keep track of all the ports in one place with the RobotMap Global Variable.