Add Teleop-like Functionality to Auton
Since Teleop.vi and Autonomous.vi never run at the same time, they can write to the same variable without a race condition. IS THIS RIGHT? This way the robot drives EXACTLY the same becasue everything controling it is exactly the same

To implement the autonomous functionality, write to more global variables and make more cases in the case structures or something, but a a lower priority level than manual

per the 2019 control system (https://wpilib.screenstepslive.com/s/currentCS/m/getting_started/l/1028812-new-for-2019), there will be a ~0.5s delay between Auton and Teleop where robots are in the DIsabled. SO, a smooth transition between the stages isn't going to work.