Auton a little more finalized
I did a little more work on the auton (because I couldn't find work anywhere else), it
should take the selected name, find the index where the name is stored, look for that same index in where the steps are stored, then move the steps to the buffer, which is finally moved to the Output.

Not 100% sure if writing to the buffer will work.

Still need appropriate values for each step.