Moved a few of the periodic task bits to Interpret Joysticks
Specifically, I have only done about half of it so far the Drive inputs and 4bar inputs. This is to decide what to do with the inputs (math and stuff) in one VI, and outputting all the motor values into global variables to use in their seperate VIs. I'm pushing it early in the case that something gets screwed up while I finish it.

Co-Authored-By: cjsmith02 <cjsmith02@users.noreply.github.com>