Done-ish
Should have a function to incorperate a climb system (2 motors) into the robot code using driver bumpers to use climb.
The max height is read using a limit switch that, once hit, causes the driver controller to rumble for 200 ms before shutting down the climb control loop.
The -ish is because i feel like there could be some edits made to the code, not fully sure if the rumble will work in its current state, but there is another system that will do the same by putting that function in a seperate while loop, I was also considering changing it so that instead of stopping the loop once the limit switch hit, it would just change the value writing to the motor from increasing to zero, allowing you to close the climb mech if needed. (Not implemented in any form yet)