Investigated Timer issue
Looks fine to me, but apparently it doesn't work. I added a little comment where the timer is, I think that the clock needs to have a connection to an outside timer using whatever weird connection thing you have at the competition, and I don't know whether or not we have that connection. A normal timer would probably work better.

Also fixed this annoying unsymmetrical wire